requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.4
django-admin-rangefilter==0.11.1
django-elasticsearch-dsl==7.3
django-extensions==3.2.3
django-filter==23.2
django-haystack==3.2.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
djangorestframework==3.14.0
drf-yasg==1.21.7
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
inflection==0.5.1
packaging==23.1
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
